git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/wc-collate-path@1405866 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/svn/mergeinfo-cmd.c
@@ -152,8 +152,8 @@ mergeinfo_diagram(svn_client__pathrev_t *yca,
       col_width[col] = 0;
       for (row = 0; row < ROWS; row++)
         {
-          if (g[row][col] && (strlen(g[row][col]) > col_width[col]))
-            col_width[col] = strlen(g[row][col]);
+          if (g[row][col] && ((int)strlen(g[row][col]) > col_width[col]))
+            col_width[col] = (int)strlen(g[row][col]);
         }
     }
